app/requirements.txt
arabic-reshaper==2.0.14
astroid==1.6.5
certifi==2018.4.16
click==6.6
configparser==3.5.0
cycler==0.10.0
Flask==1.0
Flask-SQLAlchemy==2.1
-e git://github.com/gangverk/flask-swagger.git@9e2868fbe6947e8c8cc312b12d7abe8892759773#egg=flask_swagger
future==0.16.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.8
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==0.23
marshmallow==2.15.1
marshmallow-sqlalchemy==0.12.0
matplotlib==2.2.2
mccabe==0.6.1
multidict==4.3.1
nose2==0.6.5
numpy==1.15.0
persian-wordcloud==1.3.1
Pillow==6.2.0
pylint==1.9.2
pyparsing==2.2.0
python-bidi==0.4.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.1
six==1.10.0
SQLAlchemy==1.1.2
Werkzeug==0.15.3
wordcloud==1.4.1
wrapt==1.10.11
